+++ 02-ignite-timer/src/contexts/CyclesContext.tsx
@@ -1,0 +1,110 @@
+import { createContext, ReactNode, useState } from "react";
+
+type Cycle = {
+  id: string;
+  task: string;
+  minutesAmount: number;
+  startDate: Date;
+  interruptedDate?: Date;
+  finishedDate?: Date;
+};
+
+interface ICreateCycleData {
+  task: string;
+  minutesAmount: number;
+}
+
+interface ICyclesContextType {
+  cycles: Cycle[];
+  activeCycle: Cycle | undefined;
+  activeCycleId: string | null;
+  amoundSecondsPassed: number;
+  markCurrentCyclesAsFinished: () => void;
+  setSecondsPassed: (seconds: number) => void;
+  createNewCycle: (data: ICreateCycleData) => void;
+  interruptCurrentCycle: () => void;
+}
+
+interface ICyclesContextProviderProps {
+  children: ReactNode;
+}
+
+export const CyclesContext = createContext({} as ICyclesContextType);
+
+export function CyclesContextProvider({
+  children,
+}: ICyclesContextProviderProps) {
+  const [cycles, setCycles] = useState<Cycle[]>([]);
+  const [activeCycleId, setActiveCycleId] = useState<string | null>(null);
+  const [amoundSecondsPassed, setAmoundSecondsPassed] = useState(0);
+
+  const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId);
+
+  function setSecondsPassed(seconds: number) {
+    setAmoundSecondsPassed(seconds);
+  }
+
+  function markCurrentCyclesAsFinished() {
+    setCycles((state) =>
+      state.map((cycle) => {
+        if (cycle.id === activeCycleId) {
+          return {
+            ...cycle,
+            finishedDate: new Date(),
+          };
+        }
+
+        return cycle;
+      })
+    );
+  }
+
+  function interruptCurrentCycle() {
+    setCycles((state) =>
+      state.map((cycle) => {
+        if (cycle.id === activeCycleId) {
+          return {
+            ...cycle,
+            interruptedDate: new Date(),
+          };
+        }
+
+        return cycle;
+      })
+    );
+
+    setActiveCycleId(null);
+  }
+
+  function createNewCycle(data: ICreateCycleData) {
+    const newCycle: Cycle = {
+      id: new Date().getTime().toString(),
+      task: data.task,
+      minutesAmount: data.minutesAmount,
+      startDate: new Date(),
+    };
+
+    setCycles((oldCyclesValues) => [...oldCyclesValues, newCycle]);
+    setActiveCycleId(newCycle.id);
+    setAmoundSecondsPassed(0);
+
+    // reset();
+  }
+
+  return (
+    <CyclesContext.Provider
+      value={{
+        cycles,
+        activeCycle,
+        activeCycleId,
+        amoundSecondsPassed,
+        markCurrentCyclesAsFinished,
+        setSecondsPassed,
+        createNewCycle,
+        interruptCurrentCycle,
+      }}
+    >
+      {children}
+    </CyclesContext.Provider>
+  );
+}
