+++ 02-ignite-timer/src/pages/History/index.tsx
@@ -1,9 +1,17 @@
+import { useContext } from "react";
+
+import { CyclesContext } from "../../contexts/CyclesContext";
+
 import { HistoryContainer, HistoryList, Status } from "./styles";
 
 export default function History() {
+  const { cycles } = useContext(CyclesContext);
+
   return (
     <HistoryContainer>
       <h1>Meu histórico</h1>
+
+      <pre>{JSON.stringify(cycles, null, 2)}</pre>
 
       <HistoryList>
         <table>
