+++ 02-ignite-timer/src/pages/Home/components/Countdown/index.tsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect } from "react";
 import { differenceInSeconds } from "date-fns";
 
-import { CyclesContext } from "../..";
+import { CyclesContext } from "../../../../contexts/CyclesContext";
 
 import { CountDownContainer, Separator } from "./styles";
